Move more constructor tests, explode, drop
