+++ python/cudf/cudf/tests/test_series.py
@@ -75,226 +75,6 @@ def test_series_error_equality(sr1, sr2, op):
 
 
 @pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series(["a"] * 20, index=range(0, 20)),
-        pd.Series(["b", None] * 10, index=range(0, 20), name="ASeries"),
-        pd.Series(
-            ["b", None] * 5,
-            index=pd.Index(list(range(10)), dtype="uint64"),
-            name="BSeries",
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "labels",
-    [
-        [1],
-        [0],
-        1,
-        5,
-        [5, 9],
-        pd.Index([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]),
-        pd.Index([0, 1, 2, 3, 4], dtype="float32"),
-    ],
-)
-@pytest.mark.parametrize("inplace", [True, False])
-def test_series_drop_labels(ps, labels, inplace):
-    ps = ps.copy()
-    gs = cudf.from_pandas(ps)
-
-    expected = ps.drop(labels=labels, axis=0, inplace=inplace)
-    actual = gs.drop(labels=labels, axis=0, inplace=inplace)
-
-    if inplace:
-        expected = ps
-        actual = gs
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series(["a"] * 20, index=range(0, 20)),
-        pd.Series(["b", None] * 10, index=range(0, 20), name="ASeries"),
-    ],
-)
-@pytest.mark.parametrize(
-    "index",
-    [[1], [0], 1, 5, [5, 9], pd.Index([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])],
-)
-@pytest.mark.parametrize("inplace", [True, False])
-def test_series_drop_index(ps, index, inplace):
-    ps = ps.copy()
-    gs = cudf.from_pandas(ps)
-
-    expected = ps.drop(index=index, inplace=inplace)
-    actual = gs.drop(index=index, inplace=inplace)
-
-    if inplace:
-        expected = ps
-        actual = gs
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series(
-            ["a" if i % 2 == 0 else "b" for i in range(0, 10)],
-            index=pd.MultiIndex(
-                levels=[
-                    ["lama", "cow", "falcon"],
-                    ["speed", "weight", "length"],
-                ],
-                codes=[
-                    [0, 0, 0, 1, 1, 1, 2, 2, 2, 1],
-                    [0, 1, 2, 0, 1, 2, 0, 1, 2, 1],
-                ],
-            ),
-            name="abc",
-        )
-    ],
-)
-@pytest.mark.parametrize(
-    "index,level",
-    [
-        ("cow", 0),
-        ("lama", 0),
-        ("falcon", 0),
-        ("speed", 1),
-        ("weight", 1),
-        ("length", 1),
-        (
-            "cow",
-            None,
-        ),
-        (
-            "lama",
-            None,
-        ),
-        (
-            "falcon",
-            None,
-        ),
-    ],
-)
-@pytest.mark.parametrize("inplace", [True, False])
-def test_series_drop_multiindex(ps, index, level, inplace):
-    ps = ps.copy()
-    gs = cudf.from_pandas(ps)
-
-    expected = ps.drop(index=index, inplace=inplace, level=level)
-    actual = gs.drop(index=index, inplace=inplace, level=level)
-
-    if inplace:
-        expected = ps
-        actual = gs
-
-    assert_eq(expected, actual)
-
-
-def test_series_drop_edge_inputs():
-    gs = cudf.Series([42], name="a")
-    ps = gs.to_pandas()
-
-    assert_eq(ps.drop(columns=["b"]), gs.drop(columns=["b"]))
-
-    assert_eq(ps.drop(columns="b"), gs.drop(columns="b"))
-
-    assert_exceptions_equal(
-        lfunc=ps.drop,
-        rfunc=gs.drop,
-        lfunc_args_and_kwargs=(["a"], {"columns": "a", "axis": 1}),
-        rfunc_args_and_kwargs=(["a"], {"columns": "a", "axis": 1}),
-    )
-
-    assert_exceptions_equal(
-        lfunc=ps.drop,
-        rfunc=gs.drop,
-        lfunc_args_and_kwargs=([], {}),
-        rfunc_args_and_kwargs=([], {}),
-    )
-
-    assert_exceptions_equal(
-        lfunc=ps.drop,
-        rfunc=gs.drop,
-        lfunc_args_and_kwargs=(["b"], {"axis": 1}),
-        rfunc_args_and_kwargs=(["b"], {"axis": 1}),
-    )
-
-
-def test_series_drop_raises():
-    gs = cudf.Series([10, 20, 30], index=["x", "y", "z"], name="c")
-    ps = gs.to_pandas()
-
-    assert_exceptions_equal(
-        lfunc=ps.drop,
-        rfunc=gs.drop,
-        lfunc_args_and_kwargs=(["p"],),
-        rfunc_args_and_kwargs=(["p"],),
-    )
-
-    # dtype specified mismatch
-    assert_exceptions_equal(
-        lfunc=ps.drop,
-        rfunc=gs.drop,
-        lfunc_args_and_kwargs=([3],),
-        rfunc_args_and_kwargs=([3],),
-    )
-
-    expect = ps.drop("p", errors="ignore")
-    actual = gs.drop("p", errors="ignore")
-
-    assert_eq(actual, expect)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [[[1, 2, 3], None, [4], [], [5, 6]], [1, 2, 3, 4, 5]],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-@pytest.mark.parametrize(
-    "p_index",
-    [
-        None,
-        ["ia", "ib", "ic", "id", "ie"],
-        pd.MultiIndex.from_tuples(
-            [(0, "a"), (0, "b"), (0, "c"), (1, "a"), (1, "b")]
-        ),
-    ],
-)
-def test_explode(data, ignore_index, p_index):
-    pdf = pd.Series(data, index=p_index, name="someseries")
-    gdf = cudf.from_pandas(pdf)
-
-    expect = pdf.explode(ignore_index)
-    got = gdf.explode(ignore_index)
-
-    assert_eq(expect, got, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [[1, 2, 3], [10, 20]],
-        [[1.0, 2.0, 3.0], None, [10.0, 20.0, np.nan]],
-        [[5, 6], None, [1]],
-        [None, None, None, None, None, [10, 20]],
-    ],
-)
-@pytest.mark.parametrize("klass", [cudf.Series, list, cp.array])
-def test_nested_series_from_sequence_data(data, klass):
-    actual = cudf.Series(
-        [klass(val) if val is not None else val for val in data]
-    )
-    expected = cudf.Series(data)
-    assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize(
     "data",
     [
         cp.ones(5, dtype=cp.float16),
